+++ internal/provider/incident_alert_source_resource.go
@@ -189,11 +189,13 @@ func (r *IncidentAlertSourceResource) Create(ctx context.Context, req resource.C
 		return
 	}
 
-	result, err := r.client.AlertSourcesV2CreateWithResponse(ctx, client.AlertSourcesCreatePayloadV2{
-		Name:        data.Name.ValueString(),
-		SourceType:  client.AlertSourcesCreatePayloadV2SourceType(data.SourceType.ValueString()),
-		Template:    data.Template.ToPayload(),
-		JiraOptions: data.JiraOptions.ToPayload(),
+	result, err := lockForAlertConfig(ctx, func(ctx context.Context) (*client.AlertSourcesV2CreateResponse, error) {
+		return r.client.AlertSourcesV2CreateWithResponse(ctx, client.AlertSourcesCreatePayloadV2{
+			Name:        data.Name.ValueString(),
+			SourceType:  client.AlertSourcesCreatePayloadV2SourceType(data.SourceType.ValueString()),
+			Template:    data.Template.ToPayload(),
+			JiraOptions: data.JiraOptions.ToPayload(),
+		})
 	})
 
 	if err != nil {
@@ -244,10 +246,12 @@ func (r *IncidentAlertSourceResource) Update(ctx context.Context, req resource.U
 		return
 	}
 
-	result, err := r.client.AlertSourcesV2UpdateWithResponse(ctx, data.ID.ValueString(), client.AlertSourcesUpdatePayloadV2{
-		Name:        data.Name.ValueString(),
-		Template:    data.Template.ToPayload(),
-		JiraOptions: data.JiraOptions.ToPayload(),
+	result, err := lockForAlertConfig(ctx, func(ctx context.Context) (*client.AlertSourcesV2UpdateResponse, error) {
+		return r.client.AlertSourcesV2UpdateWithResponse(ctx, data.ID.ValueString(), client.AlertSourcesUpdatePayloadV2{
+			Name:        data.Name.ValueString(),
+			Template:    data.Template.ToPayload(),
+			JiraOptions: data.JiraOptions.ToPayload(),
+		})
 	})
 
 	if err != nil {
@@ -268,7 +272,9 @@ func (r *IncidentAlertSourceResource) Delete(ctx context.Context, req resource.D
 		return
 	}
 
-	_, err := r.client.AlertSourcesV2DeleteWithResponse(ctx, data.ID.ValueString())
+	_, err := lockForAlertConfig(ctx, func(ctx context.Context) (*client.AlertSourcesV2DeleteResponse, error) {
+		return r.client.AlertSourcesV2DeleteWithResponse(ctx, data.ID.ValueString())
+	})
 	if err != nil {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to delete alert source, got error: %s", err))
 		return