+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Unreleased
+
+- Enforce consistent ordering of keys in engine literal values which are JSON objects.
+- Made `handovers` on `incident_schedule.config.rotations.*` required, as it
+  was always required by the API, but was not marked as such in the provider.
+
+## v5.9.1
+
+- Improve documentation to reflect potential values of enumerated values.
+- Improve example for `incident_alert_route` documentation.
+
 ## v5.9.0
 
 - Update the documentation for `incident_alert_source`, `incident_alert_route`, `incident_escalation_path` and `incident_schedule` to reference the 'Export' flow